asked for confirmation of vue and react

+++ installation/installation.sh
@@ -41,14 +41,34 @@ printf "\n"
 nvm install 8.10
 
 # Vue
-printf "\nInstalling Vue CLI"
-printf "\n\n"
-npm install -g @vue/cli
+
+# Asking for confimation
+printf "\nDo you want to install Vue? [y/n] -> "
+read vueAnswer
+
+yes_no $vueAnswer
+
+if [ $? == 0 ]
+then
+	printf "\nInstalling Vue CLI"
+	printf "\n\n"
+	npm install -g @vue/cli
+else
+	printf "\n\nOkay, skipping Vue"
+
+fi
+
+
+
 
 # AWS and SAM CLI
-printf "\n\nEnter your local machine username -> "
-read username
+printf "\n\nThe following is the logged in user."
+printf "\n\n"
+whoami
 printf "\n\nInstalling the AWS and SAM CLI \n\n ..."
+printf "\n\nEnter the local machine username that you wish the AWS CLI to be associated with -> "
+read username
+printf "\n\nInstalling as $username"
 printf "\n\nWhile this is running, ensure you have your AWS Access Key ID and AWS Secret Access Key ready to go. \n\n"
 sudo -u $username brew tap aws/tap
 sudo -u $username brew install awscli aws-sam-cli
@@ -63,7 +83,7 @@ aws configure
 # GitHub SSH Key Gen
 
 # Asking for confirmation
-printf "\n\nThis is the last step and is optional. \nIt will generate and copy an ssh key to your clipboard.\n\nDo you want to continue? [y/n] -> "
+printf "\n\nThis step will generate and copy an ssh key to your clipboard.\n\nDo you want to continue? [y/n] -> "
 read githubAnswer
 
 yes_no $githubAnswer
@@ -84,12 +104,32 @@ printf "\nDone!\nThe SSH key is copied to the clipboard"
 open https://github.com/settings/keys
 
 else 
+	printf "\n\nAlright, be that way. Moving on!"
+	exit 0
+fi
+
+
+# React CLI
+
+# Asking for confirmation
+printf "\n\nDo you want to install the React CLI? [y/n] -> "
+read reactAnswer
+
+yes_no $reactAnswer
+
+if [ $? == 0]
+then 
+printf "\n\nOkay, installing the React tools"
+npm install -g create-react-app
+
+else 
 	printf "\n\nAlright, be that way. All done!"
 	exit 0
 fi
 
 
-
+printf "\n\nAll done!"
+exit 0
